Fix time dependent test, upgrade karma-expect->karma-chai-plugins
karma-expect docs say:
> You should use it only if you want run tests in IE8 and lower,
> otherwise consider karma-chai-plugins, which provides more
> complete Chai assertions as well as additional plugins.

We don't need IE8, so let's switch.

The failing test uses fixed dates in April 2018 and assumes that this
date is somewhat recent, which is not true anymore. I left the fixed
date in place since data from April 2018 is now old enough to have
reached the 1-hour-rollups-are-kept-forever age, thus the test should
work as is indefinitely.